Seperated killing player from Game over loader
Seperated the code that kills the player from the gameover loader so it doesn't accidentally get forgetten. Now Kill the Player is on the dynamic sprite in PlzGameOver, so you can still test it
