+++ Assets/Scenes/Development Scenes/PlzGameOver.unity
@@ -583,6 +583,7 @@ GameObject:
   - component: {fileID: 1644484007}
   - component: {fileID: 1644484006}
   - component: {fileID: 1644484005}
+  - component: {fileID: 1644484009}
   m_Layer: 0
   m_Name: Dynamic Sprite
   m_TagString: Untagged
@@ -719,6 +720,18 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1644484009
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1644484004}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: eef5dd17ce45b984195b7ce9f3c3f580, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1001 &4958287530160409860
 PrefabInstance:
   m_ObjectHideFlags: 0
