+++ Assets/Scripts/KillThePlayers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eef5dd17ce45b984195b7ce9f3c3f580
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
