+++ Assets/Scripts/UI/LoadGameOverScreen.cs
@@ -27,22 +27,6 @@ public class LoadGameOverScreen : MonoBehaviour
     }
     
 
-    // Update is called once per frame
-    void Update()
-    {
-        DamageInfo damageInfo = new DamageInfo(100, Vector2.right, AuraType.Strike);
-
-        if (Input.GetKeyDown(KeyCode.J))
-        {
-            playersHealth[0].Damage(damageInfo);
-            Debug.Log("Player 1 Should be damaged");
-        }
-        if (Input.GetKeyDown(KeyCode.K))
-        {
-            playersHealth[1].Damage(damageInfo);
-            Debug.Log("Player 2 Should be damaged");
-        }
-    }
 
     /// <summary>
     /// Load Scene when player dies
